Mov memory test to performance tests
