+++ tests/ert/performance_tests/test_memory_usage.py
@@ -1,18 +1,25 @@
+import os
+import stat
 import tempfile
+from multiprocessing import Process
 from pathlib import Path
-from typing import List
+from textwrap import dedent
+from typing import List, Tuple
 
 import memray
 import numpy as np
 import polars
 import py
 import pytest
+import xtgeo
 
 from ert.analysis import smoother_update
 from ert.config import ErtConfig
 from ert.enkf_main import sample_prior
+from ert.mode_definitions import ENSEMBLE_SMOOTHER_MODE
 from ert.storage import open_storage
 from tests.ert.performance_tests.performance_utils import make_poly_example
+from tests.ert.ui_tests.cli.run_cli import run_cli
 
 
 @pytest.fixture
@@ -143,4 +150,102 @@ def make_summary_data(
             ).dt.cast_time_unit("ms"),
             "values": data,
         }
+    )
+
+
+@pytest.mark.limit_memory("130 MB")
+@pytest.mark.flaky(reruns=5)
+def test_field_param_memory(tmpdir):
+    with tmpdir.as_cwd():
+        # Setup is done in a subprocess so that memray does not pick up the allocations
+        p = Process(target=create_poly_with_field, args=((2000, 1000, 1), 2))
+        p.start()
+        p.join()  # this blocks until the process terminates
+
+        run_poly()
+
+
+@pytest.mark.usefixtures("set_site_config")
+def create_poly_with_field(field_dim: Tuple[int, int, int], realisations: int):
+    """
+    This replicates the poly example, only it uses FIELD parameter
+    """
+    grid_size = field_dim[0] * field_dim[1] * field_dim[2]
+    config = dedent(
+        f"""
+            NUM_REALIZATIONS {realisations}
+            OBS_CONFIG observations
+
+            FIELD MY_PARAM PARAMETER my_param.bgrdecl INIT_FILES:my_param.bgrdecl FORWARD_INIT:True
+            GRID MY_EGRID.EGRID
+
+            GEN_DATA MY_RESPONSE RESULT_FILE:gen_data_%d.out REPORT_STEPS:0 INPUT_FORMAT:ASCII
+            INSTALL_JOB poly_eval POLY_EVAL
+            FORWARD_MODEL poly_eval
+            """
+    )
+    with open("config.ert", "w", encoding="utf-8") as fh:
+        fh.writelines(config)
+
+    grid = xtgeo.create_box_grid(dimension=field_dim)
+    grid.to_file("MY_EGRID.EGRID", "egrid")
+    del grid
+
+    with open("forward_model", "w", encoding="utf-8") as f:
+        f.write(
+            f"""#!/usr/bin/env python
+import numpy as np
+import os
+import resfo
+
+if __name__ == "__main__":
+    if not os.path.exists("my_param.bgrdecl"):
+        values = np.random.standard_normal({grid_size})
+        resfo.write("my_param.bgrdecl", [("MY_PARAM", values)])
+    datas = resfo.read("my_param.bgrdecl")
+    assert datas[0][0] == "MY_PARAM"
+    a,b,c,*_ = datas[0][1]
+
+    output = [float(a) * x**2 + float(b) * x + float(c) for x in range(10)]
+    with open("gen_data_0.out", "w", encoding="utf-8") as f:
+        f.write("\\n".join(map(str, output)))
+            """
+        )
+    os.chmod(
+        "forward_model",
+        os.stat("forward_model").st_mode | stat.S_IXUSR | stat.S_IXGRP | stat.S_IXOTH,
+    )
+    with open("POLY_EVAL", "w", encoding="utf-8") as fout:
+        fout.write("EXECUTABLE forward_model")
+    with open("observations", "w", encoding="utf-8") as fout:
+        fout.write(
+            dedent(
+                """
+        GENERAL_OBSERVATION MY_OBS {
+            DATA       = MY_RESPONSE;
+            INDEX_LIST = 0,2,4,6,8;
+            RESTART    = 0;
+            OBS_FILE   = obs.txt;
+        };"""
+            )
+        )
+
+    with open("obs.txt", "w", encoding="utf-8") as fobs:
+        fobs.write(
+            dedent(
+                """
+        2.1457049781272213 0.6
+        8.769219841380755 1.4
+        12.388014786122742 3.0
+        25.600464531354252 5.4
+        42.35204755970952 8.6"""
+            )
+        )
+
+
+def run_poly():
+    run_cli(
+        ENSEMBLE_SMOOTHER_MODE,
+        "--disable-monitor",
+        "config.ert",
     )
